refactor: consolidate auth and loading page
